pyproject.toml
[tool.poetry]
name = "protein-properties"
version = "0.1.0"
description = ""
authors = ["CallMeMisterOwl <denizhasler@outlook.de>"]
readme = "README.md"
packages = [{include = "protein_properties"}]

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.24.3"
pandas = "^2.0.1"
wandb = "0.18.5"
scikit-learn = "^1.2.2"
h5py = "^3.8.0"
jupyter = "^1.0.0"
seaborn = "^0.12.2"
biotite = "1.0"
biopython = "^1.81"
tensorboard = "^2.13.0"
jupyterlab = "^4.0.2"
torchmetrics = "^0.11.4"
lightning = "^2.0.4"
jsonargparse = {version = "4.17.0", extras = ["signatures"]}
pytorch-lightning = {extras = ["extras"], version = "^2.0.4"}
captum = "^0.6.0"
statsmodels = "^0.14.0"
rcsbsearchapi = "^1.6.0"
pdbecif = "^1.5"
polars = "^1.6.0"
imbalanced-learn = "^0.12.3"
umap-learn = "^0.5.6"
tensorflow = {extras = ["and-cuda"], version = "^2.18.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
